Backend/requirements.txt
fastapi==0.109.2
uvicorn==0.23.2
pandas==2.3.2
numpy==1.26.4
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.1
